requirements.txt
streamlit==1.31.1
pandas==2.2.0
plotly==5.18.0
psycopg2-binary==2.9.9
python-dotenv
openai==1.12.0
openpyxl==3.1.2  # 엑셀 파일 처리용
